refactor(auth): simplify AuthOptions storage initialization
- Replace optional storage parameter with a non-optional default value
- Remove defaultStorage() method and related platform-specific logic
- Default storage now always uses UserDefaultsAuthStorage
- Streamline AuthOptions initializer for clarity and simplicity
